Locations map working

+++ db/migrate/20120324165316_create_locations.rb
@@ -2,8 +2,8 @@ class CreateLocations < ActiveRecord::Migration
   def self.up
     create_table :locations do |t|
       t.integer :location_id
-      t.decimal :latitude
-      t.decimal :longitude
+      t.float :latitude
+      t.float :longitude
       t.binary :map
       t.string :name
       t.string :email
